Move the Public handlers to Server.hs.
This is temporary; they will probably move elsewhere later.
